+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.4] - yyyy-mm-dd
+
+### Fixed
+
+- If a user is verified, then it set the agrement to be accepted
+
 ## [0.4.3] - yyyy-mm-dd
 
 ### Added
@@ -15,11 +21,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- On the final page, the finish remains finish
+
 ### Fixed
 
-NFT selection page related issues
-
+- NFT selection page related issues
 - If the user cancels a transaction, it still navigates to the success page, just to show an empty page
+- Various other fixes & improvements
 
 ## [0.4.2] - 2023-01-31